refactor: create AddButton
